// packages/lib/types/apiResponses.ts
export type OrderSide = 'Buy' | 'Sell' | 'Short' | 'Unknown';
export type OrderType =
  | 'Unknown'
  | 'Market order'
  | 'Limit'
  | 'StopMarket'
  | 'StopLimit'
  | 'TrailingStopMarket'
  | 'TrailingStopLimit'
  | 'BlockTrade';
export type OrderChangeReason =
  | 'Unknown'
  | 'NewInputAccepted'
  | 'NewInputRejected'
  | 'OtherRejected'
  | 'Expired'
  | 'Trade'
  | 'SystemCanceled_NoMoreMarket'
  | 'SystemCanceled_BelowMinimum'
  | 'SystemCanceled_PriceCollar'
  | 'SystemCanceled_MarginFailed'
  | 'UserModified';
export type OrderState =
  | 'Unknown'
  | 'Working'
  | 'Rejected'
  | 'Canceled'
  | 'Expired'
  | 'FullyExecuted';
// ...
export type ActivityItem = {
  id: number;
  amount: number;
  type: string;
  status: string;
  timestamp: number;
  instrumentId?: number;
  productId?: number;
  decimalPlaces?: number;
  currency?: string;
  fullName?: string;
  symbol?: string;
};

export interface OrderActivity extends ActivityItem {
  type: OrderType;
  status: OrderState;
  iconSymbols?: readonly [string, string];
  side: OrderSide;
  avgPrice: number;
  limitPrice?: number;
}